requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-auth-ldap3-ad==1.6.30
django-excel==0.0.10
django-tinymce4-lite==1.7.5
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
jsmin==2.2.2
ldap3==2.6
lml==0.0.9
openpyxl==2.6.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.7
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
texttable==1.6.2
urllib3==1.25.3
xlrd==1.2.0
xlwt==1.3.0
gunicorn==19.9.0